requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-socketio==5.5.1
browser-use==0.1.40
langchain-openai==0.3.1
python-dotenv==1.0.1
nest-asyncio==1.6.0
langchain==0.3.19
openai==1.60.2
gunicorn==23.0.0
